requirements.txt

Django==3.0.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
django-heroku==0.3.1